# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Aplikasi PKB",
        "description": "ini adalah dokumentasi API Aplikasi PKB untuk integrasi dengan sistem PKB",
        "termsOfService": "http://lpjk.pu.go.id/pkb/terms/",
        "contact": {
            "email": "admin@lpjk.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://dev.siki.pu.go.id/pkb/api/api/",
            "description": "PKB API host"
        },
        {
            "url": "http://localhost:8000/api/",
            "description": "Local API host"
        }
    ],
    "paths": {
...
        "/v1/unsur": {
            "get": {
                "tags": [
                    "Refrensi"
                ],
                "summary": "Get list of active tests",
                "description": "Returns list of active tests",
                "operationId": "get unsur list",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/v1/klasifikasi": {
            "get": {
                "tags": [
                    "Refrensi"
                ],
                "summary": "Get list of active tests",
                "description": "Returns list of active tests",
                "operationId": "get klasifikasi list",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/v1/metode": {
            "get": {
                "tags": [
                    "Refrensi"
                ],
                "summary": "Get list of active tests",
                "description": "Returns list of active tests",
                "operationId": "get metode list",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/v1/penyelenggara": {
            "get": {
                "tags": [
                    "Refrensi"
                ],
                "summary": "Get list of active tests",
                "description": "Returns list of active tests",
                "operationId": "get penyelenggara list",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    ]
}